Implement stress test on renderers.

+++ Sparky-core/src/graphics/buffers/indexbuffer.cpp
@@ -8,7 +8,7 @@ namespace sparky {
         {
             glGenBuffers(1, &m_bufferID);
             glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_bufferID);
-            glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data, GL_STATIC_DRAW);
+            glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLuint), data, GL_STATIC_DRAW);
             glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
         }
